Modify S1874: Merge with `react/no-deprecated`
